requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras pyproject.toml
#
argon2-cffi==23.1.0
    # via tumbs.app (pyproject.toml)
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.8.1
    # via django
authlib==1.2.1
    # via kinde-python-sdk
certifi==2022.12.7
    # via
    #   kinde-python-sdk
    #   requests
cffi==1.16.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==42.0.6
    # via authlib
django==5.0.4
    # via
    #   django-extensions
    #   django-redis
    #   djangorestframework
    #   drf-standardized-errors
    #   tumbs.app (pyproject.toml)
django-environ==0.11.2
    # via tumbs.app (pyproject.toml)
django-extensions==3.2.3
    # via tumbs.app (pyproject.toml)
django-ipware==7.0.1
    # via tumbs.app (pyproject.toml)
django-ordered-model==3.7.4
    # via tumbs.app (pyproject.toml)
django-redis==5.4.0
    # via tumbs.app (pyproject.toml)
django-webpack-loader==3.1.0
    # via tumbs.app (pyproject.toml)
djangorestframework==3.15.1
    # via
    #   drf-standardized-errors
    #   tumbs.app (pyproject.toml)
drf-standardized-errors==0.13.0
    # via tumbs.app (pyproject.toml)
frozendict==2.3.10
    # via kinde-python-sdk
hiredis==2.3.2
    # via tumbs.app (pyproject.toml)
idna==3.7
    # via requests
kinde-python-sdk==1.2.3
    # via tumbs.app (pyproject.toml)
pillow==10.3.0
    # via tumbs.app (pyproject.toml)
pycparser==2.22
    # via cffi
pyjwt==2.8.0
    # via kinde-python-sdk
python-dateutil==2.8.2
    # via kinde-python-sdk
python-ipware==3.0.0
    # via django-ipware
python-magic==0.4.27
    # via tumbs.app (pyproject.toml)
redis==5.0.4
    # via
    #   django-redis
    #   tumbs.app (pyproject.toml)
requests==2.31.0
    # via kinde-python-sdk
six==1.16.0
    # via python-dateutil
sqlparse==0.5.0
    # via django
typing-extensions==4.8.0
    # via kinde-python-sdk
urllib3==2.0.7
    # via
    #   kinde-python-sdk
    #   requests
whitenoise==6.6.0
    # via tumbs.app (pyproject.toml)
